+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 - Options to the `tye run` command: `--watch`, `--debug`, `--framework`
 - `tye build` and `tye push` actions to the Services tool window
+- `tye build` and `tye push` actions to the Project view window
 
 ### Changed
 - Reduce the number of supported IDE to Rider